Report memory usage when device allocation fails
TIDL API creates 2 device side heaps:
1. Parameter heap
2. Network heap

The sizes of these heaps are specified in the Configuration object, via
PARAM_HEAP_SIZE and NETWORK_HEAP_SIZE.

Existing behavior: If the heaps are not large enough, allocation on the
device triggers an assertion failure with no indication of how large the
heaps need to be for successfull allocation.

To improve the usability of the API, provide feedback to the user on the
heap sizes required to satisfy device side allocations when any
allocation fails.

Also added `-Wall -Werror` when building examples and fixed failures.

(MCT-1035)

+++ examples/segmentation/object_classes.cpp
@@ -62,7 +62,7 @@ object_class_table_t* GetObjectClassTable(std::string &config)
 
 object_class_t* GetObjectClass(object_class_table_t *table, int index)
 {
-    if (index < 0 || index >= table->num_classes)  index = table->num_classes;
+    if (index < 0 || (unsigned int)index >= table->num_classes)  index = table->num_classes;
     return & (table->classes[index]);
 }
